refactor: reorganize the code to better match the styleguide
