+++ src/api/contact/module.ts
@@ -1,0 +1,19 @@
+import { ContactSchema } from '@/components/contact/components/contact-form/schema';
+
+export const submitFormData = async (contactData: ContactSchema) => {
+  const response = await fetch('/api/contact', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(contactData),
+  });
+
+  const data = await response.json();
+
+  if (!response.ok) {
+    throw new Error(JSON.stringify(data));
+  }
+
+  return data;
+};
