+++ src/api/contact/queries.ts
@@ -1,0 +1,14 @@
+import { useMutation } from '@tanstack/react-query';
+import { submitFormData } from './module';
+
+type Props = {
+  onSuccess: () => void;
+  onError: (error: unknown) => void;
+};
+
+export const useContactMutation = ({ onSuccess, onError }: Props) =>
+  useMutation({
+    mutationFn: submitFormData,
+    onSuccess,
+    onError,
+  });
